mobilenetv2webapp1-main/mobilenetv2webapp1-main/requirements.txt
streamlit
tensorflow-cpu
numpy
Pillow
pandas
